renamed fetcher and added tests
